refactor: move two functions to utils.helpers to resolve circular dependency
